// src/app/interfaces/PrestamoSummary.ts
export interface PrestamoSummary {
    idPrestamo: number;
    fechaInicio: string;
    fechaFin: string;
    idCliente: string;
    nombres: string;
    apellidos: string;
    edad: number;
    direccion: string;
    telefono: string;
    idLibro: number;
    nombre: string;
    descripcion: string;
    categoria: string;
    autor: string;
    ano: string;
}